Move pseudo element style computation to a separate module
