+++ gaphor/core/styling/tests/test_css.py
@@ -3,9 +3,9 @@ import pytest
 from gaphor.core.styling import (
     CompiledStyleSheet,
     compile_style_sheet,
-    compute_pseudo_element_style,
 )
 from gaphor.core.styling.declarations import WhiteSpace
+from gaphor.core.styling.pseudo import compute_pseudo_element_style
 from gaphor.core.styling.tests.test_compiler import Node
